requirements.txt
PyMySQL==0.9.3
python-dotenv==0.10.1